+++ server/src/Lib/Hover/include.ts
@@ -1,4 +1,4 @@
 /*	Auto generated	*/
 
-export * from "./Mcfunction";
+export * from "../Minecraft/Mcfunction/Hover";
 export * from "./OnRequest";
