+++ server/src/Lib/Minecraft/Mcfunction/include.ts
@@ -4,3 +4,4 @@ export * from "./Completion";
 export * from "./References";
 export * from "./Semantics";
 export * from "./Signature";
+export * from "./Hover";
